+++ src/state/mod.rs
@@ -3,6 +3,7 @@ mod command;
 mod command_buffer;
 mod document;
 mod error;
+mod selection;
 mod transient;
 mod view;
 
@@ -11,5 +12,6 @@ pub use crate::state::command::*;
 pub use crate::state::command_buffer::*;
 pub use crate::state::document::*;
 pub use crate::state::error::*;
+pub use crate::state::selection::*;
 pub use crate::state::transient::*;
 pub use crate::state::view::*;
